# api-documentation.yaml
...
openapi: 3.0.0
info:
  description: This is the API for the Vehicle Cost Calculator
  version: "1.0.0"
  title: Vehicle Cost Calculator API
...
paths:
...
  /users/{username}/vehicles:
    post:
      tags:
        - Vehicles
      summary: Add a new Vehicle to the user
      description: |
        Inserts a new Vehicle for the given User into the database
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehicle'
      responses:
        '201':
          description: Vehicle created succesfully
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Vehicles
      summary: Get the vehicles of the user
      description: |
        Delivers all vehicles of the given username
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
      responses:
        '200':
          description: Vehicles of the User
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleComplete'
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{username}/vehicles/{vin}:
    get:
      tags:
        - Vehicles
      summary: Get a specific vehicle of the user
      description: |
        Delivers a specific vehicle of the given username with the given vin
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          required: true
          description: Vin of one specific vehicle
          schema:
            type: string
            example: "1P3EW65F4VV300946"
      responses:
        '200':
          description: Vehicle of the User
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleComplete'
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user or vin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Vehicles
      summary: Deletes one Vehicle
      description: |
        Deletes one Vehicle with the given vin
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          required: true
          description: Vin of one specific vehicle
          schema:
            type: string
            example: "1P3EW65F4VV300946"
      responses:
        '204':
          description: Successfully deleted
        '400':
          description: The vin does't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Vehicles
      summary: Overwrites a Vehicle
      description: |
        Overwrites a Vehicle
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          required: true
          description: Vin of one specific vehicle
          schema:
            type: string
            example: "1P3EW65F4VV300946"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehicle'
        description: Vehicle to overwrite (vin must match vin in path)
      responses:
        '201':
          description: Vin overwritten
        '400':
          description: The vin does't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{username}/vehicles/{vin}/singleCosts:
    post:
      tags:
        - Single Costs
      summary: Add a new single cost item to the Vehicle
      description: |
        Inserts a new single cost item for the given Vehicle into the database
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          schema:
            type: string
          required: true
          description: VIN of the Vehicle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SingleCostItem'
      responses:
        '201':
          description: Single cost item created succesfully
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Single Costs
      summary: Get the single cost item of the vehicle
      description: |
        Delivers all single cost items of the given vehicle
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          schema:
            type: string
          required: true
          description: VIN of the Vehicle
      responses:
        '200':
          description: Single cost items of the vehicle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SingleCostItemComplete'
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{username}/vehicles/{vin}/fuelCosts:
    post:
      tags:
        - Fuel Costs
      summary: Add a new Fuel cost item to the Vehicle
      description: |
        Inserts a new Fuel cost item for the given Vehicle into the database
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          schema:
            type: string
          required: true
          description: VIN of the Vehicle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FuelCostItem'
      responses:
        '201':
          description: Fuel cost item created succesfully
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Fuel Costs
      summary: Get the Fuel cost item of the vehicle
      description: |
        Delivers all Fuel cost items of the given vehicle
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          schema:
            type: string
          required: true
          description: VIN of the Vehicle
      responses:
        '200':
          description: Fuel cost items of the vehicle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FuelCostItemComplete'
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{username}/vehicles/{vin}/repeatingCosts:
    post:
      tags:
        - Repeating Costs
      summary: Add a new Repeating cost item to the Vehicle
      description: |
        Inserts a new Repeating cost item for the given Vehicle into the database
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          schema:
            type: string
          required: true
          description: VIN of the Vehicle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepeatingCostItem'
      responses:
        '201':
          description: Repeating cost item created succesfully
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Repeating Costs
      summary: Get the Repeating cost item of the vehicle
      description: |
        Delivers all Repeating cost items of the given vehicle
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          schema:
            type: string
          required: true
          description: VIN of the Vehicle
      responses:
        '200':
          description: Repeating cost items of the vehicle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RepeatingCostItemComplete'
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Vehicle:
      type: object
      required:
        - vin
      properties:
        vin:
          type: string
          example: "1P3EW65F4VV300946"
        year:
          type: integer
          format: int32
          example: 2001
        make:
          type: "string"
          example: "BMW"
        model:
          type: "string"
          example: "116i"
        type:
          type: "string"
          example: "Small car"
        color:
          type: "string"
          example: "Blue"
        weight:
          type: integer
          format: int32
          example: 2862
        license:
          type: string
          example: "Tü-T-7502"
    VehicleComplete:
      type: object
      required:
        - vin
        - username
      properties:
        vin:
          type: string
          example: "1P3EW65F4VV300946"
        username:
          type: string
          example: "Mustermann25"
        year:
          type: string
          format: Date
          example: "Fri Dec 03 2010 19:26:05 GMT+0100 (Mitteleuropäische Normalzeit)"
        make:
          type: "string"
          example: "BMW"
        model:
          type: "string"
          example: "116i"
        type:
          type: "string"
          example: "Small car"
        color:
          type: "string"
          example: "Blue"
        weight:
          type: integer
          format: int32
          example: 2862
        license:
          type: string
          example: "Tü-T-7502"
    FuelCostItem:
      type: object
      required:
        - date
        - price
        - volume
      properties:
        km:
          type: integer
          format: int32
          example: 90600
          description: mileage in km
        price:
          type: number
          description: price in €
          format: float
          example: 80.4
        date:
          type: integer
          format: date
          example: 122234234
          description: gives the date of the cost item in ms
        volume:
          type: number
          format: float
          description: fuel volume in l
          example: 42.5
        consumption:
          type: number
          format: float
          description: consumption of the cost item
        type:
          type: string
          example: "super"
          description: type of the fuel. diesel | super | e10 | super plus | gaz
    FuelCostItemComplete:
      type: object
      required:
        - id
        - vin
        - username
        - date
        - price
        - volume
      properties:
        id:
          type: string
          format: uuid
          example: "6c84fb90-12c4-11e1-840d-7b25c5ee775a"
        vin:
          type: string
          example: "1P3EW65F4VV300946"
        username:
          type: string
          example: "Mustermann25"
        km:
          type: integer
          format: int32
          example: 90600
          description: mileage in km
        price:
          type: number
          description: price in €
          format: float
          example: 80.4
        date:
          type: integer
          format: date
          example: 122234234
          description: gives the date of the cost item in ms
        volume:
          type: number
          format: float
          description: fuel volume in l
          example: 42.5
        consumption:
          type: number
          format: float
          description: consumption of the cost item
        type:
          type: string
          example: "super"
          description: type of the fuel. diesel | super | e10 | super plus | gaz
    SingleCostItem:
      type: object
      required:
        - date
        - price
        - type
        - name
      properties:
        km:
          type: integer
          format: int32
          example: 90600
          description: mileage in km
        price:
          type: number
          description: price in €
          format: float
          example: 80.4
        date:
          type: integer
          format: date
          example: 122234234
          description: gives the date of the cost item in ms
        type:
          type: string
          description: type of the cost item. repair | procurement | rest
          example: "repair"
        name:
          type: string
          description: name of the cost item
          example: "new weels"
        description:
          type: string
          example: "weels for the winter season"
          description: closer description for the cost item (optional)
    SingleCostItemComplete:
      type: object
      required:
        - id
        - vin
        - username
        - date
        - price
        - type
        - name
      properties:
        id:
          type: string
          format: uuid
          example: "6c84fb90-12c4-11e1-840d-7b25c5ee775a"
        vin:
          type: string
          example: "1P3EW65F4VV300946"
        username:
          type: string
          example: "Mustermann25"
        km:
          type: integer
          format: int32
          example: 90600
          description: mileage in km
        price:
          type: number
          description: price in €
          format: float
          example: 80.4
        date:
          type: integer
          format: date
          example: 122234234
          description: gives the date of the cost item in ms
        type:
          type: string
          description: type of the cost item. repair | procurement | rest
          example: "repair"
        name:
          type: string
          description: name of the cost item
          example: "new weels"
        description:
          type: string
          example: "weels for the winter season"
          description: closer description for the cost item (optional)
    RepeatingCostItem:
      type: object
      required:
        - date
        - period
        - price
        - name
      properties:
        price:
          type: number
          description: price in €
          format: float
          example: 80.4
        date:
          type: integer
          format: date

          description: the start date of the cost item in ms
        period:
          type: string
          description: "period for this cost item"
        name:
          type: string
          description: name of the cost item
          example: "insurance"
        description:
          type: string
          example: "liability insurance"
          description: closer description for the cost item (optional)
    RepeatingCostItemComplete:
      type: object
      required:
        - id
        - vin
        - username
        - date
        - period
        - price
        - name
      properties:
        id:
          type: string
          format: uuid
          example: "6c84fb90-12c4-11e1-840d-7b25c5ee775a"
        vin:
          type: string
          example: "1P3EW65F4VV300946"
        username:
          type: string
          example: "Mustermann25"
        price:
          type: number
          description: price in €
          format: float
          example: 80.4
        date:
          type: integer
          format: date

          description: the start date of the cost item in ms
        period:
          type: string
          description: "period for this cost item"
        name:
          type: string
          description: name of the cost item
          example: "insurance"
        description:
          type: string
          example: "liability insurance"
          description: closer description for the cost item (optional)
    Error:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            type: object
            required:
              - msg
              - location
            properties:
              value:
                type: string
                example: TestValue
              msg:
                type: string
                example: Invalid Value
              param:
                type: string
                example: Password
              location:
                type: string
                example: body